Flytt periodeOverlapperMedValgtDato-funksjon til nytt sted
