+++ src/frontend/utils/dato/index.ts
@@ -1,1 +1,2 @@
 export * from './dato';
+export * from './periode';
